refactor(settings): split provider delete helper
